move eth provider out of index and fix existed test
