+++ src/base_provider.js
@@ -1,0 +1,29 @@
+// Copyright © 2017-2020 Trust Wallet.
+//
+// This file is part of Trust. The full Trust copyright notice, including
+// terms governing use, modification, and redistribution, is contained in the
+// file LICENSE at the root of the source code distribution tree.
+
+"use strict";
+
+import { EventEmitter } from "events";
+
+class BaseProvider extends EventEmitter {
+  /**
+   * @private Internal js -> native message handler
+   */
+  postMessage(handler, id, data) {
+    let object = {
+      id: id,
+      name: handler,
+      object: data,
+    };
+    if (window.trustwallet.postMessage) {
+      window.trustwallet.postMessage(object);
+    } else {
+      console.error("postMessage is not available");
+    }
+  }
+}
+
+module.exports = BaseProvider;
